// generated/src/app/api-models/member-working-aged.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberWorkingAged {
  seqMembId: number;
  seqMembWaId: number;
  dateSent: Date;
  dateReceived: Date;
  memwaUserDefined1: string;
  memwaUserDefined2: string;
  memwaUserDate1: Date;
  memwaUserDate2: Date;
  surveyEffectiveDate: Date;
  surveyTermDate: Date;
  insertDatetime: Date;
  insertProcess: string;
  insertUser: string;
  updateDatetime: Date;
  updateProcess: string;
  updateUser: string;
  securityCode: string;
}